update restart and stop script and update go mod
update restart and stop script and update go mod

+++ scripts/bin/restart.sh
@@ -13,14 +13,18 @@ if [ ! -z ${pid} ];then
     kill $pid
 fi
 
-docker_go_container_name=`grep 'dockervm_container_name:' ../config/{org_id}/chainmaker.yml | tail -n1 | awk '{print $2}'`
-docker_container_lists=(`docker ps -a | grep ${docker_go_container_name} | awk '{print $1}'`)
+enable_dockervm=`grep 'enable_dockervm:' ../config/{org_id}/chainmaker.yml | awk '{print $2}'`
+if [ ${enable_dockervm} == "true" ];then
+  docker_go_container_name=`grep 'dockervm_container_name:' ../config/{org_id}/chainmaker.yml | tail -n1 | awk '{print $2}'`
+  docker_container_lists=(`docker ps -a | grep ${docker_go_container_name} | awk '{print $1}'`)
+  for container_id in ${docker_container_lists[*]}
+  do
+    docker stop ${container_id}
+    docker rm ${container_id}
+  done
+fi
 
-for container_id in ${docker_container_lists[*]}
-do
-  docker stop ${container_id}
-  docker rm ${container_id}
-done
+
 
 sleep 2
 #nohup ./chainmaker start -c ../config/{org_id}/chainmaker.yml > /dev/null 2>&1 &
